refactor file structure add modal folder to component
